change sign in component to always create a new user
The current implementation of carshare-back expects the client to always
attempt to create a new user after login. It will then sort out whether
or not an insert or update is warranted.

I now think a better solution would be to have carshare-back create a
user whenever it receives any valid operation for a firebase user that
it hasn't seen before. That way the new user can just get on with using
the system and the client can finish onboarding the new user as a
background job.